split hermes and ts-relayer test
